frontend/requirements.txt
PyYAML>=6.0.1
streamlit>=1.33.0
streamlit-antd-components>=0.3.2
streamlit-on-Hover-tabs>=1.0.1
requests>=2.31.0
numpy>=1.26.4
pandas>=2.2.2
scipy>=1.13.0
shap>=0.45.1
matplotlib>=3.8.4
mplcyberpunk>=0.7.1
seaborn>=0.13.2
scikit-learn>=1.4.2
optuna>=3.6.1
joblib>=1.4.2
plotly>=5.22.0